Breaking apart the monolith.
